requirements.txt
pyserial~=3.5
PyYAML~=6.0
parcon~=0.1.25
pyelftools~=0.26